Move docker tests to a non-Windows file
